Move metrics statistics types to new file.
